split renderer files, rename installers to recipes
