+++ packages/installer/src/components/newline.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import {Box} from 'ink'
 
-export function Newline({count = 1}) {
+export const Newline: React.FC<{count?: number}> = ({count = 1}) => {
   return <Box paddingBottom={count} />
 }
